Fix deployment CI job
This change introduces new scheme for releasing the code: the code will be
uploaded only when the tag in the format 'v_<version>' is pushed to the repo.
This tag defines the version which will be dynamically calculated in setup.py.

Deploy job is now only installing twine and running it. Packaging is performed
on every release and the resulting files are stored as artifacts for 1 week.
Version number for that package is determined as version from the closest tag
in 'v_<version>' format plus shortened commit-id.

So, from now on, to release a new version:
- create a new tag: git tag v_1.23.45
- push it to the repo: git push origin v_1.23.45

This will trigger full build with deployment task in the end